+++ CHANGELOG.md
@@ -4,7 +4,8 @@ ArduinoJson: change log
 HEAD
 ----
 
-* Fix `unsigned long` printed as `signed long` (issue #170)
+* Added custom implementation of `ftoa` (issues #266, #267, #269 and #270)
+* Fixed `unsigned long` printed as `signed long` (issue #170)
 
 v5.2.0
 ------